Fix crash because of missing PACKED_VECTOR4_ARRAY during API generation with args. (#275)
* Fix crash because of missing PACKED_VECTOR4_ARRAY during API generation with args.

Added PACKED_VECTOR4_ARRAY in cpp_compatible_variant_type and added a test that generates API with argument names for all godot classes to detect incompatibilities in the future.

* Fix crash because of missing PACKED_VECTOR4_ARRAY during API generation with args.

Added PACKED_VECTOR4_ARRAY in cpp_compatible_variant_type and added a test that generates API with argument names for all godot classes to detect incompatibilities in the future.